Another round of misc edits & fixes
Deduction (3): "exceed" should probably be "succeed"; if you're breaking to bullet points midsentence, then the bullet points don't need to be capitalized, and you should use "..." to end the pre-list statement and to start each bullet point; success by X specifies either a range ("or more/less") or explicitly says "exactly X"; original Deduction (both 0 and 2) discover the additional clues at that location (which is not necessarily your location), though that leaves "connecting location" slightly ambiguous (I think the default interpretation there would be that it meant connected to "that location").
Slip By: the timing clause is organized a little awkwardly ("when [enemy] [does thing] [additional enemy qualifiers]"). If you shift that around to start with "Play when you would be engaged by ..." then you can keep all the enemy qualifiers batched together.
Flight or Fright: Weird one. AFAICT action designators are always used at the start of an ability/effect as a way of flagging it, so I'm not sure how to deal with one that lands mid-resolution. Is the playing of FoF supposed to count as both an evade and a move action? If so, maybe just: "Evade. Move. Evade an enemy, then move, or vice versa. (This move does not provoke attacks of opportunity.)"
📘 Inspect: icons should be contributed to tests, not people. Could simplify it to "While attached, Inspect contributes its skill icons to all investigation and evasion tests against the attached card." (also: the art not completely filling the frame is a little strange)
Evidence in Blood: "while evading" -> "evasion attempt" would read a little smoother
Ithaca 37 (both L0 and L2): "than the attacked enemy" -> "than that of the attacked enemy"
Resourcefulness: Oops, this is another unelaborated "Fast" event. Also the name is a bit strange when "Resourceful" is already a staple
🟪 Master of Secrets: First ability I would rewrite entirely to: "[reaction] After you investigate, if no token is sealed here, exhaust Master of Secrets: Seal a non-[tentacle] token revealed during that test. If you seal a "+0," "+1," or [elder sign] token in this way, discover 1 clue at your location." (i.e.: simplify initial clause; move 1-token-only clause to before the colon; comma placement; comma missing)
Soporific Grimoire: parenthetical sentence is missing its period
Tattoos of the Void: reaction can drop the "You may" (it's already optional to trigger); the Forced ability having to retain memory about multiple "the other card"s is pretty messy.. maybe just check for presence of identical sealed tokens?
🟥 Gambler's Resolve: the success result works regardless of who's taking the test, but the failure result only applies when it's you. maybe "If you fail" -> "If this test fails"? (also: the angles on the die in the air look kinda wrong) (also: whitespace in the art) (p.s. I really like this design! cool variant of Sneak By / BTMO)
Weathered Oilskin: I think "successfully" is unnecessary; what if you're engaged with multiple enemies?
Last Ditch Effort: I think Last-Ditch needs a hyphen?
attoos of the Void: maybe make this unique? as futureproofing against multiple-bodyslot shennanigans
Frayed Token: does it really need a round limit when it already exhausts?; why is it unique?
Night Camp: this is expensive and slow enough that it could probably safely heal damage and horror